refactor(app): split all oversized files in modules/app below 200-line limit
Split 9 files that exceeded the 200-line rule into focused companion files:
- ai-assist-ui.ts → ai-assist-popover.ts (popover/dropdown/loading DOM builders)
- toolbar-selects.ts → toolbar-style-select.ts (style picker + paragraph spacing)
- table-cell-format.ts → table-cell-dom.ts (cell toolbar DOM builder)
- editor-page.ts → editor-page-exports.ts (export/import handlers)
- doc-list-controls.ts → doc-list-pagination.ts (pagination bar)
- global-search.ts → global-search-keyboard.ts (keyboard nav + shortcut)
- editor-ruler.ts → editor-ruler-drag.ts (tab stop + margin drag handlers)
- editor-view.ts → editor-view-collab.ts (fetchMyRole + applyPermissionMode)
- editor.ts → editor-post-init.ts (template apply, page controls, metadata track)

All original public APIs re-exported from the original files — zero import changes
required in consumers. TypeScript: zero errors. Tests: 1826 passed, 0 failed.
