+++ modules/app/internal/editor/editor-ruler.ts
@@ -2,22 +2,15 @@
 
 /**
  * Horizontal document ruler with draggable margin handles and tab stops (#217).
- * Positioned between the formatting toolbar and the editor canvas in the DOM flow.
- * State is persisted in localStorage.
+ * Drag interaction logic is in editor-ruler-drag.ts.
  */
 
 import { drawRulerTicks } from './editor-ruler-canvas.ts';
 import { safeResizeObserver } from './lifecycle.ts';
+import { attachTabDrag, attachMarginDrag, type RulerState } from './editor-ruler-drag.ts';
 
 const KEY = 'opendesk-ruler';
 const DEFAULT_MARGIN = 80; // px — matches editor padding: 5rem at 16px/rem
-
-interface RulerState {
-  visible: boolean;
-  left: number;
-  right: number;
-  tabs: number[]; // px from content-area left edge
-}
 
 function loadState(): RulerState {
   try {
@@ -99,42 +92,10 @@ export function initRuler(): void {
     tabEls.length = 0;
     for (const tx of state.tabs) {
       const el = buildTabMarker(offset + state.left + tx);
-      attachTabDrag(el, tx);
+      attachTabDrag(el, tx, state, getPaperOffset, saveState, redraw, container);
       container.appendChild(el);
       tabEls.push(el);
     }
-  }
-
-  function attachTabDrag(el: HTMLElement, tabX: number): void {
-    el.addEventListener('mousedown', (e) => {
-      e.preventDefault();
-      e.stopPropagation();
-      const startClientX = e.clientX;
-      const idx = state.tabs.indexOf(tabX);
-      const { left: offset, width: pw } = getPaperOffset();
-      const contentWidth = pw - state.left - state.right;
-
-      const onMove = (me: MouseEvent): void => {
-        const dx = me.clientX - startClientX;
-        const newX = Math.max(0, Math.min(contentWidth, tabX + dx));
-        el.style.left = (offset + state.left + newX) + 'px';
-        if (idx >= 0) state.tabs[idx] = Math.round(newX);
-      };
-
-      const onUp = (me: MouseEvent): void => {
-        document.removeEventListener('mousemove', onMove);
-        document.removeEventListener('mouseup', onUp);
-        const cr = container.getBoundingClientRect();
-        if (me.clientY < cr.top - 16 || me.clientY > cr.bottom + 16) {
-          if (idx >= 0) state.tabs.splice(idx, 1);
-        }
-        saveState(state);
-        redraw();
-      };
-
-      document.addEventListener('mousemove', onMove);
-      document.addEventListener('mouseup', onUp);
-    });
   }
 
   // Click on canvas to add a tab stop
@@ -149,36 +110,8 @@ export function initRuler(): void {
     redraw();
   });
 
-  function attachMarginDrag(handle: HTMLElement, side: 'left' | 'right'): void {
-    handle.addEventListener('mousedown', (e) => {
-      e.preventDefault();
-      const startX = e.clientX;
-      const startMargin = side === 'left' ? state.left : state.right;
-      const { width: pw } = getPaperOffset();
-
-      const onMove = (me: MouseEvent): void => {
-        const dx = me.clientX - startX;
-        const delta = side === 'left' ? dx : -dx;
-        const newM = Math.max(16, Math.min(pw / 2 - 32, startMargin + delta));
-        if (side === 'left') state.left = Math.round(newM);
-        else state.right = Math.round(newM);
-        applyMargins(paper, state);
-        redraw();
-      };
-
-      const onUp = (): void => {
-        document.removeEventListener('mousemove', onMove);
-        document.removeEventListener('mouseup', onUp);
-        saveState(state);
-      };
-
-      document.addEventListener('mousemove', onMove);
-      document.addEventListener('mouseup', onUp);
-    });
-  }
-
-  attachMarginDrag(leftHandle, 'left');
-  attachMarginDrag(rightHandle, 'right');
+  attachMarginDrag(leftHandle, 'left', state, getPaperOffset, (s) => applyMargins(paper, s), saveState, redraw);
+  attachMarginDrag(rightHandle, 'right', state, getPaperOffset, (s) => applyMargins(paper, s), saveState, redraw);
 
   function setVisible(v: boolean): void {
     state.visible = v;
