+++ modules/app/internal/editor/table-cell-format.ts
@@ -2,6 +2,8 @@
 import type { Editor } from '@tiptap/core';
 import type { Transaction } from 'prosemirror-state';
 
+export { buildCellFormatSection } from './table-cell-dom.ts';
+
 interface CellColor {
   label: string;
   color: string;
@@ -165,59 +167,4 @@ export function setCellTextAlign(
   }
 
   if (changed) view.dispatch(tr);
-}
-
-/** Build the cell formatting section DOM and append it to the container. */
-export function buildCellFormatSection(
-  container: HTMLElement,
-  editor: Editor,
-): void {
-  // Separator
-  const sep1 = document.createElement('div');
-  sep1.className = 'table-toolbar-cell-section';
-
-  const label = document.createElement('span');
-  label.className = 'table-toolbar-label';
-  label.textContent = 'Cell:';
-  sep1.appendChild(label);
-
-  // Color swatches
-  for (const { label: colorLabel, color, style } of CELL_COLORS) {
-    const btn = document.createElement('button');
-    btn.className = 'table-toolbar-color-swatch';
-    btn.title = colorLabel;
-    btn.setAttribute('aria-label', `Cell background: ${colorLabel}`);
-    btn.setAttribute('style', style);
-    btn.addEventListener('click', (e) => {
-      e.preventDefault();
-      setCellBackground(editor, color);
-    });
-    sep1.appendChild(btn);
-  }
-
-  container.appendChild(sep1);
-
-  // Alignment section
-  const sep2 = document.createElement('div');
-  sep2.className = 'table-toolbar-cell-section';
-
-  const alignLabel = document.createElement('span');
-  alignLabel.className = 'table-toolbar-label';
-  alignLabel.textContent = 'Align:';
-  sep2.appendChild(alignLabel);
-
-  for (const { label: alignLabel2, align, icon } of CELL_ALIGN_BUTTONS) {
-    const btn = document.createElement('button');
-    btn.className = 'toolbar-btn table-toolbar-btn';
-    btn.title = alignLabel2;
-    btn.setAttribute('aria-label', alignLabel2);
-    btn.textContent = icon;
-    btn.addEventListener('click', (e) => {
-      e.preventDefault();
-      setCellTextAlign(editor, align);
-    });
-    sep2.appendChild(btn);
-  }
-
-  container.appendChild(sep2);
 }
